% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_connection.r
\name{get_connection}
\alias{get_connection}
\title{Get FGS Database Connection}
\usage{
get_connection(cname = "boreholes")
}
\arguments{
\item{cname}{(chr) A character vector of length 1 with the connection name}
}
\value{
An RPostgreSQL database connection object
}
\description{
\code{get_connection} Read FGS connection parameters. Used to connect to internal FGS data sources.
}
\details{
This function reads a standard CSV list of connections distributed internally at the FGS
These connections provide the detail for reader access to internal FGS databases.
You must have an active network connection to floridadep.net in order to use this function.
Use show_all_connections() to get a list of available connections.
}
\examples{
library(rfgs)
library(RPostgreSQL)
borehole_db <- get_connection()

#list all tables in database
dbListTables(borehole_db)

#list all fields in table
dbListFields(borehole_db, "well_index")

#get the index table
boreholes <- dbGetQuery(borehole_db, "SELECT * FROM well_index")
str(boreholes)

#get a count of wells by county
boreholes <- dbGetQuery(borehole_db, "SELECT county, count(wnumber) AS cnt FROM well_index GROUP BY county ORDER BY cnt DESC;")

#there are case problems with the county field, i.e. entries for 'Levy' and 'LEVY'. Use UPPER to fix.
boreholes <- dbGetQuery(borehole_db, "SELECT UPPER(county) as county, count(wnumber) AS cnt FROM well_index GROUP BY county ORDER BY cnt DESC;")

#get the average totaldepth by county
boreholes <- dbGetQuery(borehole_db, "SELECT UPPER(county) AS county, avg(wnumber) AS mean FROM well_index GROUP BY county ORDER BY mean DESC;")

#  !!! ALWAYS DISCONNECT FROM DATABASES WHEN YOU ARE DONE  !!! #
dbDisconnect(borehole_db)
}
\seealso{
Other connection.functions: \code{\link{show_all_connections}}
}

